requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
langchain-groq>=0.1.3
